colocate redirects logic with routes
